+++ packages/planum/src/input-url/input.styles.ts
@@ -129,6 +129,7 @@ export const inputCss = css(subheaderCss, {
   color: '$textDark',
   fontWeight: '$regular',
   outline: 'none',
+  opacity: 1,
 
   '&::placeholder': {
     font: 'inherit',