+++ packages/planum-phone/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @uvodohq/planum-phone
 
+## 0.3.1
+
+### Patch Changes
+
+- - fix number field disabled issue on iphones
+
+## 0.3.0
+
+### Minor Changes
+
+- - disabled input color issue on iphones
+  - password check strength translations
+
 ## 0.2.1
 
 ### Patch Changes